requirements.txt
requests~=2.31.0
fastapi~=0.104.1
starlette~=0.27.0
uvicorn=0.16.0
python-docx